+++ packages/hedera-wallet-snap/packages/snap/src/facades/StakeHbarFacade.ts
@@ -134,7 +134,12 @@ export class StakeHbarFacade {
 
       const dialogParamsForStakeHbar: DialogParams = {
         type: 'confirmation',
-        content: await SnapUtils.generateCommonPanel(origin, panelToShow),
+        content: await SnapUtils.generateCommonPanel(
+          origin,
+          network,
+          mirrorNodeUrl,
+          panelToShow,
+        ),
       };
       const confirmed = await SnapUtils.snapDialog(dialogParamsForStakeHbar);
       if (!confirmed) {