+++ packages/hedera-wallet-snap/packages/snap/src/facades/account/GetAccountInfoFacade.ts
@@ -133,7 +133,12 @@ export class GetAccountInfoFacade {
 
         const dialogParamsForHederaAccountId: DialogParams = {
           type: 'confirmation',
-          content: await SnapUtils.generateCommonPanel(origin, panelToShow),
+          content: await SnapUtils.generateCommonPanel(
+            origin,
+            network,
+            mirrorNodeUrl,
+            panelToShow,
+          ),
         };
         const confirmed = await SnapUtils.snapDialog(
           dialogParamsForHederaAccountId,