+++ packages/hedera-wallet-snap/packages/snap/src/facades/allowance/ApproveAllowanceFacade.ts
@@ -132,7 +132,12 @@ export class ApproveAllowanceFacade {
 
       const dialogParamsForApproveAllowance: DialogParams = {
         type: 'confirmation',
-        content: await SnapUtils.generateCommonPanel(origin, panelToShow),
+        content: await SnapUtils.generateCommonPanel(
+          origin,
+          network,
+          mirrorNodeUrl,
+          panelToShow,
+        ),
       };
       const confirmed = await SnapUtils.snapDialog(
         dialogParamsForApproveAllowance,