+++ packages/hedera-wallet-snap/packages/snap/src/facades/allowance/DeleteAllowanceFacade.ts
@@ -100,7 +100,12 @@ export class DeleteAllowanceFacade {
 
       const dialogParamsForDeleteAllowance: DialogParams = {
         type: 'confirmation',
-        content: await SnapUtils.generateCommonPanel(origin, panelToShow),
+        content: await SnapUtils.generateCommonPanel(
+          origin,
+          network,
+          mirrorNodeUrl,
+          panelToShow,
+        ),
       };
       const confirmed = await SnapUtils.snapDialog(
         dialogParamsForDeleteAllowance,