+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/CreateTokenFacade.ts
@@ -49,7 +49,8 @@ export class CreateTokenFacade {
   ): Promise<TxReceipt> {
     const { origin, state } = walletSnapParams;
 
-    const { hederaEvmAddress, hederaAccountId, network } = state.currentAccount;
+    const { hederaEvmAddress, hederaAccountId, network, mirrorNodeUrl } =
+      state.currentAccount;
 
     const { privateKey, publicKey, curve } =
       state.accountState[hederaEvmAddress][network].keyStore;
@@ -190,7 +191,12 @@ export class CreateTokenFacade {
 
       const dialogParams: DialogParams = {
         type: 'confirmation',
-        content: await SnapUtils.generateCommonPanel(origin, panelToShow),
+        content: await SnapUtils.generateCommonPanel(
+          origin,
+          network,
+          mirrorNodeUrl,
+          panelToShow,
+        ),
       };
       const confirmed = await SnapUtils.snapDialog(dialogParams);
       if (!confirmed) {