+++ packages/hedera-wallet-snap/packages/snap/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Removed the `header` parameter from `signMessage` API
 - Utilized the `copyable` method to display text passed from Dapps as recommended by MetaMask which ignores Markdown and other special characters
+- Now showing the current network and mirror node URL at the top of all the Snap dialog boxes
 
 ### :bug: Bug Fixes